Pom file and a[[lication files configured for mvc

+++ src/main/java/com/example/demo/controller/ResumeController.java
@@ -25,7 +25,7 @@ public class ResumeController {
     }
 
     @GetMapping("/{id}")
-    public ResumeDto findResumeById(@PathVariable Integer id) {
+    public ResumeDto findResumeById(@PathVariable Long id) {
         return resumeService.findResumeById(id);
     }
